Moved parseDate to /utils
